Fix swipe right crash on pending

+++ app/src/main/java/com/example/emafoods/feature/pending/domain/usecase/DeletePendingFoodUseCase.kt
@@ -1,6 +1,7 @@
 package com.example.emafoods.feature.pending.domain.usecase
 
 import com.example.emafoods.core.data.models.Food
+import com.example.emafoods.core.domain.models.State
 import com.example.emafoods.core.domain.repository.FoodRepository
 import javax.inject.Inject
 
@@ -8,6 +9,10 @@ class DeletePendingFoodUseCase @Inject constructor(
     private val foodRepository: FoodRepository
 ) {
 
-    suspend fun execute(food: Food) =
-        foodRepository.deletePendingFood(food)
+    suspend fun execute(food: Food) : State<Food> {
+        if(food.id.isEmpty()) {
+            return State.failed("Nu mai exista reteta in asteptare")
+        }
+        return foodRepository.deletePendingFood(food)
+    }
 }
